bitrate history queue is per receiver.
