+++ internal/capture/buckets/queue.go
@@ -1,6 +1,7 @@
 package buckets
 
 import (
+	"math"
 	"sync"
 	"time"
 )
@@ -59,4 +60,29 @@ func (q *queue) avgLastN(n int) int {
 		sum += v.bitrate
 	}
 	return sum / n
+}
+
+func (q *queue) normaliseBitrate(currentBitrate int) int {
+	avgBitrate := float64(q.avg())
+	histLen := float64(q.len())
+
+	q.push(elem{
+		bitrate: currentBitrate,
+		created: time.Now(),
+	})
+
+	if avgBitrate == 0 || histLen == 0 || currentBitrate == 0 {
+		return currentBitrate
+	}
+
+	lastN := int(math.Floor(float64(currentBitrate) / avgBitrate * histLen))
+	if lastN > q.len() {
+		lastN = q.len()
+	}
+
+	if lastN == 0 {
+		return currentBitrate
+	}
+
+	return q.avgLastN(lastN)
 }
